Carousel slide effect added but still not responsive
