+++ about.html
@@ -179,7 +179,7 @@
             </article>
             <article class="postcard light blue">
               <a class="postcard__img_link">
-                <img class="postcard__img" src="photo/team-doctor2.png" alt="Image Title" />
+                <img class="postcard__img" src="photo/eyeseeu3" alt="Image Title" />
               </a>
               <div class="postcard__text t-dark">
                 <h1 class="postcard__title green">Dr. Val Everson Sienes
