--- blog-add-blogpost.html
@@ -1,95 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-  <meta charset="UTF-8">
-  <meta http-equiv="X-UA-Compatible" content="IE=edge">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title>Add Blog</title>
-  <!-- BOOTSTRAP -->
-  <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.2.2/dist/css/bootstrap.min.css" rel="stylesheet"
-    integrity="sha384-Zenh87qX5JnK2Jl0vWa8Ck2rdkQ2Bzep5IDxbcnCeuOxjzrPF/et3URy9Bv1WTRi" crossorigin="anonymous" />
-  <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap-icons@1.10.1/font/bootstrap-icons.css" />
-  <!-- FONT AWESOME -->
-  <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.min.css" />
-  <!-- CSS -->
-  <link rel="stylesheet" href="style.css" />
-  <!-- JS -->
-
-  <!-- FIREBASE -->
-  <script defer type="module" src="js/blog-add-blogpost.js"></script>
-
-  <style>
-
-    #blog-form {
-      margin-left: 6.25em;
-      margin-right: 6.25em;
-    }
-  </style>
-</head>
-
-<body>
-
-  <form id="blog-form" class="p-5 mt-5 bg-light border border-1 shadow-lg">
-    <div class="text-center mb-3">
-      <h3>Add Blog Post</h3>
-    </div>
-    <div class="mb-3 col-md-5">
-      <label for="blogId" class="form-label">Blog Id</label>
-      <input type="number" class="form-control" id="blogId" required>
-    </div>
-    <div class="mb-3 col-md-12">
-      <label for="title" class="form-label">Title</label>
-      <input type="text" class="form-control" id="title" required>
-    </div>
-    <!-- <div class="mb-3">
-      <label for="uploadImage" class="form-label">Upload Cover Image</label>
-      <input class="form-control" type="file" id="uploadImage">
-    </div>     -->
-    <div>
-      <label for="uploadImage" class="form-label">Upload Cover Image</label>
-    </div>   
-    <div class="input-group mb-3">
-      <input type="file" class="form-control" id="uploadImage" aria-describedby="upload" aria-label="Upload">
-      <button class="btn btn-outline-secondary" type="button" id="btnUpload">Upload</button>
-    </div>
-
-    <div class="mb-3 col-md-12">
-      <label for="uploadedImageURL" class="form-label">Cover Image URL</label>
-      <input type="text" class="form-control" id="uploadedImageURL" required>
-    </div>
-    <div class="mb-3 col-md-6">
-      <label for="pubdate" class="form-label">Publication Date</label>
-      <input type="datetime-local" class="form-control" id="pubdate" required>
-    </div>
-    <div class="mb-3 col-md-5 ">
-      <label for="author" class="form-label">Author</label>
-      <input type="text" class="form-control" id="author" required>
-    </div>
-    <div class="mb-3 col-md-12">
-      <label for="contentPreview" class="form-label">Content Preview</label>
-      <textarea class="form-control" id="contentPreview" rows="5" required></textarea>
-    </div>
-    <div class="mb-3 col-md-12">
-      <label for="content" class="form-label">Blog Content</label>
-      <textarea class="form-control" id="content" rows="20" required></textarea>
-    </div>
-    <button type="submit" value="submit" class="btn btn-primary col-md-3">Submit</button>
-
-    <!-- <div class="mb-3 col-md-5 ">
-      <label for="tags" class="form-label">Category or Tags (if more than one, separate by comma)</label>
-      <input type="text" class="form-control" id="tags">
-    </div> -->
-  </form>
-
-  <div class="d-flex justify-content-center">
-    <h1> . </h1>
-  </div>
-
-
-  <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.2.2/dist/js/bootstrap.bundle.min.js"
-    integrity="sha384-OERcA2EqjJCMA+/3y+gxIOqMEjwtxJY7qPCqsdltbNJuaOe923+mo//f6V8Qbsw3"
-    crossorigin="anonymous"></script>
-</body>
-
-</html>
