--- blog-right-panel.html
@@ -1,123 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-  <meta charset="UTF-8">
-  <meta http-equiv="X-UA-Compatible" content="IE=edge">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title>Blog Right Panel</title>
-  <!-- BOOTSTRAP -->
-  <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.2.2/dist/css/bootstrap.min.css" rel="stylesheet"
-    integrity="sha384-Zenh87qX5JnK2Jl0vWa8Ck2rdkQ2Bzep5IDxbcnCeuOxjzrPF/et3URy9Bv1WTRi" crossorigin="anonymous" />
-  <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap-icons@1.10.1/font/bootstrap-icons.css" />
-  <!-- FONT AWESOME -->
-  <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.min.css" />
-  <!-- CSS -->
-  <link rel="stylesheet" href="style.css" />
-  <link rel="stylesheet" href="css/blog-right-panel.css" />
-  <!-- JS -->
-
-  <!-- FIREBASE -->
-  <!-- <script defer type="module" src="js/blog-right-panel.js"></script> -->
-
-  <!-- MOCKAPI -->
-  <script defer type="module" src="js/blog-right-panel-nofirebase.js"></script>
-
-</head>
-
-<body>
-
-  <!-- NAVBAR SECTION ------------------------------------------------------->
-  <nav class="navbar navbar-expand-lg bg-light fixed-top">
-    <div class="container">
-      <a class="custom-navbarbrand fs-4" href="index.html">
-        <img src="photo/main-logo&favicon.png" alt="Logo" width="35" height="45"
-          class="d-inline-block align-text-center" />
-        <span>EyeSeeU </span>Optical
-      </a>
-      <button class="navbar-toggler" type="button" data-bs-toggle="collapse" data-bs-target="#navbarSupportedContent"
-        aria-controls="navbarSupportedContent" aria-expanded="false" aria-label="Toggle navigation">
-        <span class="navbar-toggler-icon"></span>
-      </button>
-      <div class="collapse navbar-collapse" id="navbarSupportedContent">
-        <ul class="navbar-nav ms-auto mb-2 mb-lg-0">
-          <li class="nav-item">
-            <a class="nav-link" aria-current="page" href="index.html">HOME</a>
-          </li>
-          <li class="nav-item">
-            <a class="nav-link" href="location.html">LOCATION</a>
-          </li>
-          <li class="nav-item dropdown">
-            <a class="nav-link dropdown-toggle" href="about.html" role="button" data-bs-toggle="dropdown"
-              aria-expanded="false">ABOUT
-            </a>
-            <ul class="dropdown-menu">
-              <li>
-                <a class="dropdown-item" href="about.html#aboutus">About Us</a>
-              </li>
-              <li>
-                <a class="dropdown-item" href="about.html#team">Meet the Team</a>
-              </li>
-            </ul>
-          </li>
-          <li class="nav-item dropdown">
-            <a class="nav-link dropdown-toggle" href="services.html" role="button" data-bs-toggle="dropdown"
-              aria-expanded="false">SERVICES
-            </a>
-            <ul class="dropdown-menu">
-              <li>
-                <a class="dropdown-item" href="services.html">Services</a>
-              </li>
-              <li>
-                <a class="dropdown-item" href="ishihara.html">Ishihara Test</a>
-              </li>
-            </ul>
-          </li>
-          <li class="nav-item">
-            <a class="nav-link" href="blog.html">BLOGS</a>
-          </li>
-          <li class="nav-item">
-            <a class="nav-link" href="reviews.html">REVIEWS</a>
-          </li>
-          <li class="nav-item">
-            <a class="nav-link" href="contact.html">CONTACT</a>
-          </li>
-        </ul>
-      </div>
-    </div>
-  </nav>
-  <!-- END OF NAVBAR ------------------------------------------------------------------>
-
-
-
-  <main class="d-flex flex-wrap gap-3">
-    <section class="article">
-      <div class="blog-header">
-        <h2>Stay Informed with Blogs from EyeSeeU Optical</h2>
-        <p>Stay informed by reading the helpful blogs from EyeSeeU Optical and learn more about various eye health topics and vision care. </p>
-      </div>
-      <div id="blog" class="d-grid gap-3">
-        <!-- FETCH BLOGS, DISPLAY CARDS HERE -->
-      </div>
-      <button id="morePageBtn" class="btn btn-primary my-3">View more blogs</button>
-    </section>
-    
-    <aside class="sticky d-flex justify-content-end">
-      <div class="right-sidebar border-0 p-2 shadow-lg rounded-3 bg-primary">
-        <h5 class="text-center mt-3 text-light">Blogs</h5>
-        <ul id="blogMonth" class="list-group border-0 p-2">
-          <!-- FETCH MONTHS, DISPLAY MONTHS HERE -->
-        </ul>
-      </div>
-    </aside>
-  </main>
-
-  <footer>Footer</footer>
-
-  <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.2.2/dist/js/bootstrap.bundle.min.js"
-    integrity="sha384-OERcA2EqjJCMA+/3y+gxIOqMEjwtxJY7qPCqsdltbNJuaOe923+mo//f6V8Qbsw3"
-    crossorigin="anonymous"></script>
-</body>
-
-</html>
-
