+++ blog.html
@@ -17,14 +17,10 @@
   <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.2.0/css/all.min.css" />
   <!-- CSS -->
   <link rel="stylesheet" href="style.css" />
-  <link rel="stylesheet" href="css/blog-right-panel.css" />
-  <!-- JS -->
-
+  <link rel="stylesheet" href="css/blog.css" />
  <!-- MOCKAPI -->
-<script defer type="module" src="js/blog-right-panel.js"></script>
+<script defer type="module" src="js/blog.js"></script>
 
-  
-  <!-- <script defer type="module" src="js/blog-right-panel-nofirebase.js"></script> -->
 
 </head>
