--- blogpost-draft.html
@@ -1,61 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-  <meta charset="UTF-8">
-  <meta http-equiv="X-UA-Compatible" content="IE=edge">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title>Blog Right Panel</title>
-  <!-- BOOTSTRAP -->
-  <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.2.2/dist/css/bootstrap.min.css" rel="stylesheet"
-    integrity="sha384-Zenh87qX5JnK2Jl0vWa8Ck2rdkQ2Bzep5IDxbcnCeuOxjzrPF/et3URy9Bv1WTRi" crossorigin="anonymous" />
-  <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap-icons@1.10.1/font/bootstrap-icons.css" />
-  <!-- FONT AWESOME -->
-  <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.min.css" />
-  <!-- CSS -->
-  <link rel="stylesheet" href="style.css" />
-  <link rel="stylesheet" href="css/blog-right-panel.css" />
-  <!-- JS -->
-
-  <!-- FIREBASE -->
-  <!-- <script defer type="module" src="js/blog-right-panel.js"></script> -->
-
-  <!-- MOCKAPI -->
-  <script defer type="module" src="js/blogpost-draft-nofirebase.js"></script>
-
-</head>
-
-<body>
-
-  <div class="m-5 p-5 bg-light text-center">HEADER</div>      
-  
-  <main class="bg-warning py-5" >
-    <section class="">      
-      <div class="bg-light">
-        <div id="blog-title" class="py-5 text-center">          
-          <!-- FETCH BLOG TITLE, AUTHOR, DATE PUBLISHED AND DISPLAY HERE -->
-          <!-- <h1>BLOG TITLE</h1>
-          <p>AUTHOR</p>
-          <p><small>DATE PUBLSHED</small></p> -->
-        </div>
-        <div id="blog-cover-image" class="pb-5">BLOG COVER IMAGE
-          <!-- FETCH IMAGE AND DISPLAY HERE -->
-        </div>
-        <div id="blog-content" class="pb-5">BLOG CONTENT
-          <!-- FETCH ARTICLE ND DISPLAY HERE -->
-        </div>
-      </div>
-      
-    </section>    
-    <div class="m-5 p-5 bg-light text-center">CAROUSEL OF BLOG PREVIEW</div>      
-  </main>
-
-  <footer class="text-dark">Footer</footer>
-
-  <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.2.2/dist/js/bootstrap.bundle.min.js"
-    integrity="sha384-OERcA2EqjJCMA+/3y+gxIOqMEjwtxJY7qPCqsdltbNJuaOe923+mo//f6V8Qbsw3"
-    crossorigin="anonymous"></script>
-</body>
-
-</html>
-
