+++ blogpost.html
@@ -17,15 +17,11 @@
   <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.2.0/css/all.min.css" />
   <!-- CSS -->
   <link rel="stylesheet" href="style.css" />
-  <link rel="stylesheet" href="css/blog-right-panel.css" />
+  <link rel="stylesheet" href="css/blog.css" />
   <!-- JS -->
 
-  <!-- FIREBASE -->
-  <!-- <script defer type="module" src="js/blog-right-panel.js"></script> -->
-
-  <!-- MOCKAPI -->
   <script defer type="module" src="js/blogpost.js"></script>
-  <!-- <script defer type="module" src="js/carouselcard.js"></script> -->
+
