+++ css/aboutus.css
@@ -30,7 +30,7 @@ body {
 
 .custom-navbarbrand {
   color: black;
-  margin-right: 1rem;
+  margin-right: 6.25em;
   font-size: 1.75rem;
   text-decoration: none;
   white-space: nowrap;
