--- css/blog-right-panel.css
@@ -1,90 +1,0 @@
-body {
-  margin-top: 10%;
-  scroll-behavior: smooth;
-  background: url(https://www.freepnglogos.com/uploads/circle-png/circle-blue-shadow-png-clipart-13.png) no-repeat right;
-  background-attachment: fixed;
-
-}
-
-html {
-  font-family: Arial, Helvetica, sans-serif;
-  height: 200vh;
-}
-
-
-
-
-
-
-
-main {
-  max-width: 90%;
-  margin-top: 80px;
-  margin-left: 6.25em;
-  margin-right: 6.25em;
-  margin-bottom: 1.25em;
-}
-
-.article {
-  flex-grow: 999;
-  flex-basis: 0;
-}
-
-.blog-header {
-  margin-bottom: 3rem;
-}
-
-.grid div {
-  height: 10rem;
-  /* center the text */
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-.sticky {
-  height: 100%;
-  width: 10rem;
-  flex-grow: 1;
-  flex-basis: 200px;
-  position: sticky;
-  top: 5rem;
-  cursor: pointer;
-}
-
-.card-image {
-  height: 100%;
-  width: auto;
-  object-fit: cover;
-  object-position: center;
-}
-
-.blog-image {
-  height: 70vh;
-  width:100%;
-  object-fit: cover;
-  object-position: center;
-}
-
-
-
-
-
-@media screen and (max-width: 756px) {
-
-  main {
-    margin-left: 1.5em;
-    margin-right: 1.5em;
-    /* flex-flow: column-reverse; */
-  }
-
-  /* TODO: if smaller size, make flex column reverese and no more sticky */
-
-  /* .sticky {
-    position: relative;
-  } */
-
-}
-
-
-
