+++ css/blog.css
@@ -1,59 +1,67 @@
-@import url('https://fonts.googleapis.com/css?family=Poppins:200,300,400,500,600,700,800,900&display=swap');
-/* @import url('https://fonts.googleapis.com/css2?family=Montserrat:wght@300;400;600&family=Roboto:wght@400;500;700&display=swap'); */
-
-* {
-    margin: 0;
-    padding: 0;
-    box-sizing: border-box;
-    font-family: 'Poppins', sans-serif;
-    /* font-family: 'Montserrat', sans-serif; */
-
-}
-
 body {
     margin-top: 10%;
-    height: 100%;
     scroll-behavior: smooth;
-    background: url(https://www.freepnglogos.com/uploads/circle-png/circle-blue-shadow-png-clipart-13.png) no-repeat right;
+    background: url(https://www.freeiconspng.com/uploads/blue-circle-png-30.png) no-repeat right;
     background-attachment: fixed;
-
-}
-
-
-
-/* STYLES NAVBAR ****************************************************************/
-/* Lustria font family ng favicon */
-.nav-link {
-    color: black;
-}
-
-.nav-link:hover {
-    color: #009EFF;
-    transform: scale(1.03);
-    transition: all .4s ease;
-}
-
-.custom-navbarbrand {
-    color: black;
-    /* padding-top: .3125rem;
-  padding-bottom: .3125rem; */
-    margin-right: 1rem;
-    font-size: 1.75rem;
-    text-decoration: none;
-    white-space: nowrap;
-}
-
-.custom-navbarbrand:hover {
-    color: black;
-    transform: scale(1.03);
-    transition: all .3s ease;
-}
-
-.navbar {
-    box-shadow: 0 2px 4px 1px rgba(0, 0, 0, .3);
-}
-
-
+  
+  }
+  
+  html {
+    font-family: Arial, Helvetica, sans-serif;
+    height: 200vh;
+  }
+  
+  
+  main {
+    max-width: 90%;
+    margin-top: 80px;
+    margin-left: 6.25em;
+    margin-right: 6.25em;
+    margin-bottom: 1.25em;
+  }
+  
+  .article {
+    flex-grow: 999;
+    flex-basis: 0;
+  }
+  
+  .blog-header {
+    margin-bottom: 3rem;
+  }
+  
+  .grid div {
+    height: 10rem;
+    /* center the text */
+    display: flex;
+    align-items: center;
+    justify-content: center;
+  }
+  
+  .sticky {
+    height: 100%;
+    width: 10rem;
+    flex-grow: 1;
+    flex-basis: 200px;
+    position: sticky;
+    top: 5rem;
+    cursor: pointer;
+  }
+  
+  .card-image {
+    height: 100%;
+    width: auto;
+    object-fit: cover;
+    object-position: center;
+  }
+  
+  .blog-image {
+    height: 70vh;
+    width:100%;
+    object-fit: cover;
+    object-position: center;
+  }
+  
+  
 .card-right {
 
 float: right;
@@ -69,3 +77,22 @@ width: 18rem;
 
     transform: scale(1.01);
 }
+  
+  
+@media screen and (max-width: 756px) {
+  
+  main {
+    margin-left: 1.5em;
+    margin-right: 1.5em;
+  
+  }
+
+}
+  
+@media (max-width: 576px) {
+  .d-none.d-sm-block {
+    display: none;
+  }
+}
+  
+
