+++ js/blogpost.js
@@ -51,7 +51,7 @@ function displayBlogPost() {
 }
 
 
-const API_URL = 'https://638eb1de9cbdb0dbe31294ba.mockapi.io/blogsnew?sortBy=Publish_Date&order=desc';
+const API_URL = 'https://638eb1de9cbdb0dbe31294ba.mockapi.io/blogsnew';
 
 async function getBlogs() {
   try {
@@ -63,70 +63,65 @@ async function getBlogs() {
   }
 }
 
-// Displays the Carousel Content
 async function displayBlogs() {
   const blogs = await getBlogs();
   const carouselInner = document.querySelector('.carousel-inner');
   const carouselIndicators = document.querySelector('.carousel-indicators');
   let counter = 0;
 
- 
+  blogs.forEach((blog, index) => {
+    if (counter >= 3) {
+      return;
+    }
+    counter++;
 
-//Limit Carousel to 3 slides 
-blogs.forEach((blog, index) => {
-  if (counter >= 3) {
-    return;
-  }
-  counter++;
+    const div = document.createElement('div');
+    div.classList.add('carousel-item');
 
-  const div = document.createElement('div');
-  div.classList.add('carousel-item');
+    if (index === 0) {
+      div.classList.add('active');
+    }
 
-  if (index === 0) {
-    div.classList.add('active');
-  }
-
-  // Carousel Content
-  div.innerHTML = `
-    <div class="row">
+    div.innerHTML = `
+    <div class="row d-flex justify-content-evenly">
       <div class="col-4">
-        <div class="card" style="width: 18rem;">
-          <img src="${blog.Cover_Image}" class="card-img-top" alt="...">
+        <div class="card border border-info" style="width: 18rem;">
           <div class="card-body">
-            <h5 class="card-title">${blog.Title}</h5>
+          <img src="${blogs[index].Cover_Image}" class="card-img-top" alt="...">
+            <h5 class="card-title">${blogs[index].Title}</h5>
             <p class="card-text"></p>
-            <a href="blogpost.html?blog=${blog.id}" class="btn btn-primary">Read More</a>
+            <a href="blogpost.html?blog=${blogs[index].id}" class="btn btn-info">Read More</a>
           </div>
         </div>
       </div>
       <div class="col-4">
-        <div class="card" style="width: 18rem;">
-          <img src="${blog.Cover_Image}" class="card-img-top" alt="...">
+        <div class="card border border-info " style="width: 18rem;">
           <div class="card-body">
-            <h5 class="card-title">${blog.Title}</h5>
+          <img src="${blogs[index + 3].Cover_Image}" class="card-img-top" alt="...">
+            <h5 class="card-title">${blogs[index + 3].Title}</h5>
             <p class="card-text"></p>
-            <a href="blogpost.html?blog=${blog.id}" class="btn btn-primary">Read More</a>
+            <a href="blogpost.html?blog=${blogs[index + 3].id}" class="btn btn-info">Read More</a>
           </div>
         </div>
       </div>
       <div class="col-4">
-        <div class="card" style="width: 18rem;">
-          <img src="${blog.Cover_Image}" class="card-img-top" alt="...">
+        <div class="card border border-info" style="width: 18rem;">
           <div class="card-body">
-            <h5 class="card-title">${blog.Title}</h5>
+          <img src="${blogs[index + 6].Cover_Image}" class="card-img-top" alt="...">
+            <h5 class="card-title">${blogs[index + 6].Title}</h5>
             <p class="card-text"></p>
-            <a href="blogpost.html?blog=${blog.id}" class="btn btn-primary">Read More</a>
+            <a href="blogpost.html?blog=${blogs[index + 6].id}" class="btn btn-info">Read More</a>
           </div>
         </div>
       </div>
     </div>
+    </div>
   `;
-  carouselInner.appendChild(div);
+    carouselInner.appendChild(div);
 
-  // Carousel Effects
-  const indicator = document.createElement('li');
-  indicator.setAttribute('data-bs-target', '#carouselExampleDark');
-  indicator.setAttribute('data-bs-slide-to', index);
+    const indicator = document.createElement('li');
+    indicator.setAttribute('data-bs-target', '#carouselExampleDark');
+    indicator.setAttribute('data-bs-slide-to', index);
 
     if (index === 0) {
       indicator.classList.add('active');
@@ -136,7 +131,6 @@ blogs.forEach((blog, index) => {
 }
 
 displayBlogs();
-
